refactor(calendar): extract event management logic into useCalendarEvents hook
Move event-related functions (load, save, create, update, delete) from CalendarMain component to a reusable useCalendarEvents hook to improve code maintainability and reusability
